requirements.txt
streamlit==1.32.0
google-generativeai>=0.4.1
pandas==2.2.1
fpdf==1.7.2
python-dotenv==1.0.0